Remove SUBSCRIBE_PENDING and UNSUBSCRIBE_PENDING states.
They don't add much value at the moment.

The original idea was that these states would distinguish between intent to
(un)subscribe before and after the (un)subscription request was sent to the
hub. This would help with later retrying the operations (e.g., if the subscribe
API call fails, then we can later retry them for all subscriptions that were in
state SUBSCRIBING but not SUBSCRIBE_PENDING). But this distinction is probably
not all that relevant, since it's easily imaginable that a subscription gets
stuck in SUBSCRIBE_PENDING state when a verification callback fails. So a
recovery job would probably want to retry both SUBSCRIBING and SUBSCRIBE_PENDING
states, making the distinction meaningless.